refactor: setup tailwind ui
