+++ packages/app/main/decrypt/styles/tailwind.css
@@ -1,0 +1,10 @@
+@tailwind base;
+
+body {
+    padding: 0;
+    box-sizing: border-box;
+}
+
+
+@tailwind components;
+@tailwind utilities;
