+++ packages/app/main/index.tsx
@@ -1,6 +1,6 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import { App } from './app.js'
+import { App } from './App.js'
 
 const root = document.createElement('main')
 document.body.appendChild(root)
