+++ packages/app/main/plugin-host/enable.ts
@@ -1,4 +1,6 @@
 import './register.js'
+
+import { noop } from 'lodash-es'
 import { Emitter } from '@servie/events'
 import { BooleanPreference, CurrentSNSNetwork, startPluginSNSAdaptor } from '@masknet/plugin-infra/content-script'
 import {
@@ -13,7 +15,6 @@ import type { UnboundedRegistry } from '@dimensiondev/holoflows-kit'
 import { ThemeMode, FontSize } from '@masknet/web3-shared-base'
 import { addListener } from '../message.js'
 import { worker } from '../rpc.js'
-import { noop } from 'lodash-es'
 
 // #region Setup storage
 const inMemoryStorage = createKVStorageHost(
