+++ packages/app/package.json
@@ -6,8 +6,22 @@
         ".": null
     },
     "scripts": {
-        "start": "npx webpack serve"
+        "start": "npx webpack serve",
+        "build:css": "cross-env TAILWIND_MODE=build NODE_ENV=production postcss main/decrypt/styles/tailwind.css -o main/decrypt/styles/index.css",
+        "watch:css": "cross-env TAILWIND_MODE=watch NODE_ENV=development postcss main/decrypt/styles/tailwind.css -o main/decrypt/styles/index.css --watch"
     },
+    "browserslist": {
+        "production": [
+          ">0.2%",
+          "not dead",
+          "not op_mini all"
+        ],
+        "development": [
+          "last 1 chrome version",
+          "last 1 firefox version",
+          "last 1 safari version"
+        ]
+      },
     "dependencies": {
         "@masknet/base": "workspace:^",
         "@masknet/encryption": "workspace:^",
@@ -32,7 +46,21 @@
         "crypto-browserify": "^3.12.0",
         "idb": "^7.1.1",
         "safari-14-idb-fix": "^3.0.0",
-        "stream-browserify": "^3.0.0"
+        "stream-browserify": "^3.0.0",
+        "@headlessui/react": "^1.7.11",
+        "@heroicons/react": "^2.0.16",
+        "@tailwindcss/aspect-ratio": "^0.4.2",
+        "@tailwindcss/forms": "^0.5.3",
+        "@tailwindcss/line-clamp": "^0.4.2",
+        "@tailwindcss/typography": "^0.5.9",
+        "@tailwindcss/container-queries": "^0.1.0",
+        "autoprefixer": "^10.4.13",
+        "cross-env": "^7.0.3",
+        "cssnano": "^5.1.15",
+        "postcss": "^8.4.21",
+        "postcss-cli": "^10.1.0",
+        "purgecss": "^5.0.0",
+        "tailwindcss": "^3.2.7"
     },
     "devDependencies": {
         "@pmmmwh/react-refresh-webpack-plugin": "^0.5.10",
