+++ packages/app/postcss.config.cjs
@@ -1,0 +1,7 @@
+module.exports = {
+    plugins: {
+      tailwindcss: {},
+      autoprefixer: {}
+    }
+  };
+
