+++ packages/app/tailwind.config.cjs
@@ -1,0 +1,36 @@
+const defaultTheme = require('tailwindcss/defaultTheme')
+
+module.exports = {
+  // Uncomment the line below to enable the experimental Just-in-Time ("JIT") mode.
+  // https://tailwindcss.com/docs/just-in-time-mode
+  // mode: "jit",
+  theme: {
+    extend: {
+      fontFamily: {
+        sans: ['Inter var', ...defaultTheme.fontFamily.sans],
+      },
+    },
+  },
+  variants: {},
+  plugins: [
+    require('@tailwindcss/typography'),
+    require('@tailwindcss/forms'),
+    require('@tailwindcss/line-clamp'),
+    require('@tailwindcss/aspect-ratio'),
+  ],
+  purge: {
+    // Filenames to scan for classes
+    content: [
+      "./main/**/*.html",
+      "./main/**/*.js",
+      "./main/**/*.jsx",
+      "./main/**/*.ts",
+      "./main/**/*.tsx",
+    ],
+    // Options passed to PurgeCSS
+    options: {
+      // Whitelist specific selectors by name
+      // whitelist: [],
+    }
+  }
+};
